Clean up handling of CDH parcel hashes
The parcel provisioning script now ensures that the checksum downloaded
along with the parcel is saved to both a .sha1 file and a .sha file, to
avoid errors from Cloudera Manager that the checksum file has an error.
The file used to perform verification via sha1sum is created separately
as a .shacheck file with the expected format for sha1sum, and is removed
following verification.

The check file now specifically uses two spaces between the checksum and
the file name, which some stringent implementations of sha1sum require.