Accommodate single partitions when rewriting partition table
The commands sent by rewrite_root_disk_partition_table.sh to fdisk
assumed that there is more than one partition on the root device. The
extra command for selecting for first partition when moving the
beginning of data in that partition would therefore set the start sector
to 1, which is typically wrong. The script now counts the number of
partitions on the root device and does not issue explicit partition
selection commands to fdisk when there is only one partition.