Use sfdisk for rewriting MBR partitions
The rewrite_root_disk_partition_table.sh script now uses sfdisk to non-
interactively alter the partition table for the root device when
resizing non-GPT partitions, instead of the interactive fdisk. The
inconsistent behavior of fdisk prompting when only a single partition is
present on the device made the prior technique of feeding commands via
echo to fdisk unreliable.