Adding normalization step configuration to instance templates.
Granular normalization configuration has been added to instance templates.
This will allow specific normalization steps to be disabled or enabled on a
per instance basis.

This is done by replacing the normalizeInstance boolean in the instance
template with a NormalizationConfig structure which contains configuration
settings for each individual step, similar to the Director level configs that
perform the same operation.

Director level normalization configs are used only if there are no
NormalizationConfigs specified in an instance template. Older instance
templates will be interpreted as if normalizeInstance = true is an instance
template with an empty set of NormalizationConfigs. This will allow for global
overrides.  Newer instance templates may be specified with specific
normalization configs per instance template.